+++ sketch-core/src/androidInstrumentedTest/kotlin/com/github/panpf/sketch/core/android/test/fetch/ResourceUriFetcherTest.kt
@@ -18,79 +18,44 @@ package com.github.panpf.sketch.core.android.test.fetch
 import android.content.res.Resources
 import android.net.Uri
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.panpf.sketch.test.utils.getTestContext
-import com.github.panpf.sketch.source.DrawableDataSource
-import com.github.panpf.sketch.source.ResourceDataSource
 import com.github.panpf.sketch.fetch.ResourceUriFetcher
 import com.github.panpf.sketch.fetch.newResourceUri
 import com.github.panpf.sketch.request.ImageRequest
+import com.github.panpf.sketch.source.DrawableDataSource
+import com.github.panpf.sketch.source.ResourceDataSource
 import com.github.panpf.sketch.test.singleton.getTestContextAndSketch
-import com.github.panpf.tools4j.test.ktx.assertNoThrow
-import com.github.panpf.tools4j.test.ktx.assertThrow
 import kotlinx.coroutines.runBlocking
-import org.junit.Assert
-import org.junit.Test
 import org.junit.runner.RunWith
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
+import kotlin.test.assertNotEquals
+import kotlin.test.assertNotNull
+import kotlin.test.assertNull
+import kotlin.test.assertTrue
 
 @RunWith(AndroidJUnit4::class)
 class ResourceUriFetcherTest {
 
     @Test
     fun testNewUri() {
-        val context = getTestContext()
-
-        Assert.assertEquals(
-            "android.resource://resource?resType=drawable&resName=ic_launcher",
-            newResourceUri("drawable", "ic_launcher")
+        assertEquals(
+            expected = "android.resource:///drawable/ic_launcher",
+            actual = newResourceUri("drawable", "ic_launcher")
         )
-        Assert.assertEquals(
-            "android.resource://resource?resType=drawable1&resName=ic_launcher1",
-            newResourceUri("drawable1", "ic_launcher1")
+        assertEquals(
+            expected = "android.resource:///55345",
+            actual = newResourceUri(55345)
         )
 
-        Assert.assertEquals(
-            "android.resource://resource?resId=55345",
-            newResourceUri(55345)
-        )
-        Assert.assertEquals(
-            "android.resource://resource?resId=55346",
-            newResourceUri(55346)
+        assertEquals(
+            expected = "android.resource://com.github.panpf.sketch.sample/drawable/ic_launcher",
+            actual = newResourceUri("com.github.panpf.sketch.sample", "drawable", "ic_launcher")
         )
 
-        Assert.assertEquals(
-            "android.resource://resource?packageName=testPackage&resType=drawable&resName=ic_launcher",
-            newResourceUri("testPackage", "drawable", "ic_launcher")
-        )
-        Assert.assertEquals(
-            "android.resource://resource?packageName=testPackage1&resType=drawable1&resName=ic_launcher1",
-            newResourceUri("testPackage1", "drawable1", "ic_launcher1")
-        )
-
-        Assert.assertEquals(
-            "android.resource://resource?packageName=testPackage&resId=55345",
-            newResourceUri("testPackage", 55345)
-        )
-        Assert.assertEquals(
-            "android.resource://resource?packageName=testPackage1&resId=55346",
-            newResourceUri("testPackage1", 55346)
-        )
-
-        Assert.assertEquals(
-            "android.resource://resource?packageName=${context.packageName}&resType=drawable&resName=ic_launcher",
-            context.newResourceUri("drawable", "ic_launcher")
-        )
-        Assert.assertEquals(
-            "android.resource://resource?packageName=${context.packageName}&resType=drawable1&resName=ic_launcher1",
-            context.newResourceUri("drawable1", "ic_launcher1")
-        )
-
-        Assert.assertEquals(
-            "android.resource://resource?packageName=${context.packageName}&resId=55345",
-            context.newResourceUri(55345)
-        )
-        Assert.assertEquals(
-            "android.resource://resource?packageName=${context.packageName}&resId=55346",
-            context.newResourceUri(55346)
+        assertEquals(
+            expected = "android.resource://com.github.panpf.sketch.sample/55345",
+            actual = newResourceUri("com.github.panpf.sketch.sample", 55345)
         )
     }
 
@@ -99,34 +64,31 @@ class ResourceUriFetcherTest {
         val (context, sketch) = getTestContextAndSketch()
         val testAppPackage = context.packageName
         val fetcherFactory = ResourceUriFetcher.Factory()
-        val androidResUriByName = newResourceUri(testAppPackage, "drawable", "ic_launcher")
-        val androidResUriById = newResourceUri(
-            testAppPackage,
-            com.github.panpf.sketch.test.utils.core.R.drawable.ic_launcher
+        val resourceUriByName = newResourceUri(
+            resType = "drawable",
+            resName = "ic_launcher"
+        )
+        val resourceUriById = newResourceUri(
+            resId = com.github.panpf.sketch.test.utils.core.R.drawable.ic_launcher
+        )
+        val resourceUriByName2 = newResourceUri(
+            packageName = testAppPackage,
+            resType = "drawable",
+            resName = "ic_launcher"
+        )
+        val resourceUriById2 = newResourceUri(
+            packageName = testAppPackage,
+            resId = com.github.panpf.sketch.test.utils.core.R.drawable.ic_launcher
         )
         val httpUri = "http://sample.com/sample.jpg"
         val contentUri = "content://sample_app/sample"
 
-        fetcherFactory.create(sketch, ImageRequest(context, androidResUriByName))!!.apply {
-            Assert.assertEquals(androidResUriByName, this.contentUri.toString())
-        }
-        fetcherFactory.create(sketch, ImageRequest(context, androidResUriById))!!.apply {
-            Assert.assertEquals(androidResUriById, this.contentUri.toString())
-        }
-        fetcherFactory.create(sketch, ImageRequest(context, androidResUriByName))!!.apply {
-            Assert.assertEquals(androidResUriByName, this.contentUri.toString())
-        }
-        fetcherFactory.create(sketch, ImageRequest(context, androidResUriById))!!.apply {
-            Assert.assertEquals(androidResUriById, this.contentUri.toString())
-        }
-        fetcherFactory.create(sketch, ImageRequest(context, androidResUriByName))!!.apply {
-            Assert.assertEquals(androidResUriByName, this.contentUri.toString())
-        }
-        fetcherFactory.create(sketch, ImageRequest(context, androidResUriById))!!.apply {
-            Assert.assertEquals(androidResUriById, this.contentUri.toString())
-        }
-        Assert.assertNull(fetcherFactory.create(sketch, ImageRequest(context, httpUri)))
-        Assert.assertNull(fetcherFactory.create(sketch, ImageRequest(context, contentUri)))
+        assertNotNull(fetcherFactory.create(sketch, ImageRequest(context, resourceUriByName)))
+        assertNotNull(fetcherFactory.create(sketch, ImageRequest(context, resourceUriById)))
+        assertNotNull(fetcherFactory.create(sketch, ImageRequest(context, resourceUriByName2)))
+        assertNotNull(fetcherFactory.create(sketch, ImageRequest(context, resourceUriById2)))
+        assertNull(fetcherFactory.create(sketch, ImageRequest(context, httpUri)))
+        assertNull(fetcherFactory.create(sketch, ImageRequest(context, contentUri)))
     }
 
     @Test
@@ -134,122 +96,104 @@ class ResourceUriFetcherTest {
         val element1 = ResourceUriFetcher.Factory()
         val element11 = ResourceUriFetcher.Factory()
 
-        Assert.assertEquals(element1, element1)
-        Assert.assertEquals(element1, element11)
+        assertEquals(element1, element1)
+        assertEquals(element1, element11)
 
-        Assert.assertNotEquals(element1, Any())
-        Assert.assertNotEquals(element1, null)
+        assertNotEquals(element1, Any())
+        assertNotEquals(element1, null as ResourceUriFetcher.Factory?)
 
-        Assert.assertEquals(element1.hashCode(), element1.hashCode())
-        Assert.assertEquals(element1.hashCode(), element11.hashCode())
+        assertEquals(element1.hashCode(), element1.hashCode())
+        assertEquals(element1.hashCode(), element11.hashCode())
     }
 
     @Test
     fun testFetch() {
         val (context, sketch) = getTestContextAndSketch()
+        val resId = com.github.panpf.sketch.test.utils.core.R.drawable.ic_launcher
 
-        assertNoThrow {
-            runBlocking {
-                newResourceUri("drawable", "ic_launcher").let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-        assertNoThrow {
-            runBlocking {
-                newResourceUri(com.github.panpf.sketch.test.utils.core.R.drawable.ic_launcher).let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-
-        assertNoThrow {
-            runBlocking {
-                newResourceUri(context.packageName, "drawable", "ic_launcher").let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-        assertNoThrow {
-            runBlocking {
-                newResourceUri(
-                    context.packageName,
-                    com.github.panpf.sketch.test.utils.core.R.drawable.ic_launcher
-                ).let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-
-        assertNoThrow {
-            runBlocking {
-                context.newResourceUri("drawable", "ic_launcher").let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-        assertNoThrow {
-            runBlocking {
-                context.newResourceUri(com.github.panpf.sketch.test.utils.core.R.drawable.ic_launcher)
-                    .let {
-                        ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                    }.fetch()
-            }.getOrThrow()
-        }
-
-        assertThrow(Resources.NotFoundException::class) {
-            runBlocking {
-                "${ResourceUriFetcher.SCHEME}://".let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-
-        assertThrow(Resources.NotFoundException::class) {
-            runBlocking {
-                "${ResourceUriFetcher.SCHEME}://resource?packageName=fakePackageName".let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-
-        assertThrow(Resources.NotFoundException::class) {
-            runBlocking {
-                "${ResourceUriFetcher.SCHEME}://resource?packageName=${context.packageName}&resId=errorResId".let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-
-        assertThrow(Resources.NotFoundException::class) {
-            runBlocking {
-                "${ResourceUriFetcher.SCHEME}://resource?packageName=${context.packageName}&resType=drawable&resName=34&error".let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-
-        assertThrow(Resources.NotFoundException::class) {
-            runBlocking {
-                "${ResourceUriFetcher.SCHEME}://resource?packageName=${context.packageName}&resType=drawable&resName=0".let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow()
-        }
-
-        Assert.assertTrue(
-            runBlocking {
-                newResourceUri(com.github.panpf.sketch.test.utils.core.R.drawable.ic_launcher).let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow().dataSource is DrawableDataSource
+        assertTrue(
+            newResourceUri("drawable", "ic_launcher")
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow().dataSource is DrawableDataSource
         )
-        Assert.assertTrue(
-            runBlocking {
-                newResourceUri(com.github.panpf.sketch.test.utils.core.R.raw.sample).let {
-                    ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it))
-                }.fetch()
-            }.getOrThrow().dataSource is ResourceDataSource
+        assertTrue(
+            newResourceUri(resId)
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow().dataSource is DrawableDataSource
         )
+
+        assertTrue(
+            newResourceUri(context.packageName, "drawable", "ic_launcher")
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow().dataSource is DrawableDataSource
+        )
+        assertTrue(
+            newResourceUri(context.packageName, resId)
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow().dataSource is DrawableDataSource
+        )
+
+        assertTrue(
+            newResourceUri(com.github.panpf.sketch.test.utils.core.R.raw.sample)
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow().dataSource is ResourceDataSource
+        )
+
+        assertFailsWith(Resources.NotFoundException::class) {
+            "${ResourceUriFetcher.SCHEME}://"
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow()
+        }
+        assertFailsWith(Resources.NotFoundException::class) {
+            "${ResourceUriFetcher.SCHEME}://fakePackageName"
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow()
+        }
+
+        assertFailsWith(NumberFormatException::class) {
+            "${ResourceUriFetcher.SCHEME}:///errorResId"
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow()
+        }
+        assertFailsWith(NumberFormatException::class) {
+            "${ResourceUriFetcher.SCHEME}://${context.packageName}/errorResId"
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow()
+        }
+
+        assertFailsWith(Resources.NotFoundException::class) {
+            newResourceUri("drawable1", "ic_launcher")
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow()
+        }
+        assertFailsWith(Resources.NotFoundException::class) {
+            newResourceUri("drawable", "ic_launcher1")
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow()
+        }
+
+        assertFailsWith(Resources.NotFoundException::class) {
+            "${ResourceUriFetcher.SCHEME}:///drawable/ic_launcher/error"
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow()
+        }
+        assertFailsWith(Resources.NotFoundException::class) {
+            "${ResourceUriFetcher.SCHEME}://${context.packageName}/drawable/ic_launcher/error"
+                .let { ResourceUriFetcher(sketch, ImageRequest(context, it), Uri.parse(it)) }
+                .let { runBlocking { it.fetch() } }
+                .getOrThrow()
+        }
     }
 }